+++ src/models/enums.ts
@@ -1,5 +1,5 @@
 export enum StandardSettings {
-  PreferredColors = 'preferredColors',
+  FavoriteColors = 'favoriteColors',
   ElementAdjustments = 'elementAdjustments',
   KeepBadgeColor = 'keepBadgeColor',
   KeepForegroundColor = 'keepForegroundColor'
@@ -15,13 +15,13 @@ export type AllSettings = StandardSettings | AffectedSettings;
 
 export enum Commands {
   resetColors = 'peacock.resetColors',
-  saveColor = 'peacock.saveColor',
+  saveColorToFavorites = 'peacock.saveColorToFavorites',
   enterColor = 'peacock.enterColor',
   changeColorToRandom = 'peacock.changeColorToRandom',
   changeColorToVueGreen = 'peacock.changeColorToVueGreen',
   changeColorToAngularRed = 'peacock.changeColorToAngularRed',
   changeColorToReactBlue = 'peacock.changeColorToReactBlue',
-  changeColorToPreferred = 'peacock.changeColorToPreferred'
+  changeColorToFavorite = 'peacock.changeColorToFavorite'
 }
 
 export enum ElementNames {